Step 4: Add view
Note that this has to go in `src/main/resources/templates/home.html`
which the tutorial does not say.

This is what we will see now when we run our application and open the
page.

Now run `mvn spring-boot:run` then on `http://localhost:8081` we should
see a webpage with `Hello! Welcome to Bootstrap Spring Boot`